+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ## Changed
 - [[#234](https://github.com/rust-vmm/kvm-ioctls/issues/234)] vcpu: export
-reg_size as a public method.
+  reg\_size as a public method.
 - [[#243](https://github.com/rust-vmm/kvm-ioctls/pull/243)] derived the `Copy`
   trait for `IoEventAddress` and `NoDatamatch`.
 - [[#242](https://github.com/rust-vmm/kvm-ioctls/pull/242)] x86: add support