+++ CHANGELOG.md
@@ -1,12 +1,18 @@
 # Changelog
 ## [Unreleased]
 
+## [0.7.0]
+
+### Changed
 - API change in the bindings from upstream kernel changes:
-* system_event has been made into a new union
-- The x86 module has been renamed to x86_64 for consistency (matches the kernel
+  * system\_event has been made into a new union
+- The x86 module has been renamed to x86\_64 for consistency (matches the kernel
   architecture directory name)
-- Dropped "x86" (32-bit) x86 support
 - Added all features to the generated docs.rs documentation.
+
+### Removed
+
+- Dropped "x86" (32-bit) x86 support
 
 ## [0.6.0]